+++ lib/utils/consts.dart
@@ -13,7 +13,6 @@ const enableDashboardsPageFlag = 'enable_dashboards_page';
 const enableDailyOsPageFlag = 'enable_daily_os_page';
 const enableEventsFlag = 'enable_events';
 const enableSessionRatingsFlag = 'enable_session_ratings';
-const enableSyncActorFlag = 'enable_sync_actor';
 const enableProjectsFlag = 'enable_projects';
 const enableEmbeddingsFlag = 'enable_embeddings';
 const enableVectorSearchFlag = 'enable_vector_search';